pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fre-cli"
description = "Command Line Interface for FRE commands"
readme = "README.md"
requires-python = "==3.11.*" # loosen this TODO
license = "Apache-2.0"
authors = [
    {name = "MSD Workflow Team", email="oar.gfdl.workflow@noaa.gov"}
]
keywords = [
    "gfdl",
    "workflow",
    "fre",
    "fms",
]
# Some of these are conda dependencies (e.g., cdo, cmor)
# We should find out a better way to indicate those instead
# of listing them here.
# This package also requires some conda packages in the
# noaa-gfdl channel (e.g., analysis_scripts, catalogbuilder)
dependencies = [
        'analysis_scripts',
        'catalogbuilder',
        'cdo',
        'cftime',
        'click',
        'cmor',
        'cylc-flow',
        'cylc-rose',
        'jsonschema',
        'metomi-rose',
        'netCDF4',
        'numpy',
        'pyyaml',
        'xarray',
]

classifiers = [
    "Environment :: Console",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]

dynamic = [
    "version",
]

[tool.setuptools.dynamic]
version = {attr = "fre.__version__"}

[project.scripts]
fre = 'fre.fre:fre'

[project.optional-dependencies]
docs = [
    "sphinx",
    "renku-sphinx-theme",
    "sphinx-rtd-theme",
]

test = [
    "pylint",
    "pytest",
]
